Update views.py
Fixed the error where it wouldn't let you on the site if you had not logged in yet
Check if the user is authenticated: I added a check around get_next_unwatched_movie to ensure it is only called when the user is authenticated. If request.user is an AnonymousUser, get_next_unwatched_movie will not be called, preventing the error from occurring.

Handling case for anonymous users: If the user is not authenticated, the current_movie will remain None, and no movie will be displayed. This avoids passing request.user (which is an AnonymousUser) into the function and causing issues when looking for a movie to display.